==> Notebooks/01_Datos_Raw.ipynb <==
# Exploracion Inicial

import pandas as pd
import numpy as np

## Leer archivos raw

cur = pd.read_excel('../data/raw/Cursadas-Alumnos-SIGA-v2.xlsx')
dat = pd.read_excel('../data/raw/Datos-Alumnos-SIGA-v2.xlsx')
fin = pd.read_excel('../data/raw/Finales-Alumnos-SIGA-v2.xlsx')

print('shape cur:' ,cur.shape)
print('shape dat:' ,dat.shape)
print('shape fin:' ,fin.shape)

cur.head()

dat.head()

fin.head()

# cuantos alumnos en cada file
print(cur.codigo_alumno.nunique())
print(dat.codigo_alumno.nunique())
print(fin.codigo_alumno.nunique())

cur.columns

fin.columns

dat.columns

## Asignar tipos de Datos

# dat astype with dictionary
dat = dat.astype({
    'codigo_alumno': str, 
    'pais': str, 
    'localidad': str, 
    'provincia': str, 
    'titulo_secundario': str, 
    'estado_civil': str, 
    'sexo': str, 
    'año_nacimiento': int, 
    'año_ingreso': int, 
    'codigo_propuesta': str
    })

# asignar nulos
dat = dat.replace('nan', np.nan)

# astype with dictionary
cur = cur.astype({
    'codigo_alumno': str,
    'curso': str,
    'codigo_materia': str,
    'departamento': str,
    'modalidad': str,
    'turno': str,
    'ciclo_lectivo': int,
    'tipo_de_aprobacion': str,
    'cantidad_recursada_regular': int,
    'descripción_recursada_regular': str,
    'cantidad_recursada_libre': int,
    'descripción_de_recursada_libre': str,
    })

# asignar nulos
cur = cur.replace('nan', np.nan)

fin = fin.astype({
    'codigo_alumno': str,
    'codigo_materia': str,
    'anio_academico': float,
    'nota': str, # por ahora porque mezcla numeros y 'aprobados'
    'aprobado': str,
    'promociono': str,
    'año_cursada': float,
})

# asignar nulos
fin = fin.replace('nan', np.nan)

## Reduccion de Dataset

df = pd.merge(cur, fin, how = 'inner', on = 'codigo_alumno')
df.shape

df.codigo_alumno.nunique()

df = pd.merge(df, dat, how= 'inner', on = 'codigo_alumno')
df.shape

df.codigo_alumno.nunique()

df = df.codigo_alumno.unique()

len(df)

df = pd.DataFrame(df)

df.columns = ['codigo_alumno']

df

datdf = pd.merge(df, dat, how= 'inner', on = 'codigo_alumno')
datdf.shape

curdf = pd.merge(df, cur, how= 'inner', on = 'codigo_alumno')
curdf.shape

findf = pd.merge(df, fin, how= 'inner', on = 'codigo_alumno')
findf.shape

### Guardar dataset reducidos

datdf.to_csv('../data/interim/datdf.csv', index = False)
curdf.to_csv('../data/interim/curdf.csv', index = False)
findf.to_csv('../data/interim/findf.csv', index = False)

datdf.to_pickle('../data/interim/datdf.pkl')
curdf.to_pickle('../data/interim/curdf.pkl')
findf.to_pickle('../data/interim/findf.pkl')
