==> Notebooks/02_Analisis_Inicial.ipynb <==
# Exploracion Inicial

import pandas as pd
import numpy as np

# curdf = pd.read_csv('../data/interim/curdf.csv')
# datdf = pd.read_csv('../data/interim/datdf.csv')
# findf = pd.read_csv('../data/interim/findf.csv')

curdf = pd.read_pickle('../data/interim/curdf.pkl')
datdf = pd.read_pickle('../data/interim/datdf.pkl')
findf = pd.read_pickle('../data/interim/findf.pkl')


print('cur shape: ',curdf.shape)
print('dat shape: ',datdf.shape)
print('fin shape: ',findf.shape)

curdf.columns

findf.columns

datdf.columns

# Analisis Dataset Datos

## Datos Alumnos

datdf.head()

from ydata_profiling import ProfileReport

datdfpr = ProfileReport(datdf.drop(['codigo_alumno'], axis=1), title="Datos Alumnos - Profiling Report")

datdfpr.to_file("reportes/Alumnos_ProfileReport.html")

### Criterios basados en el análsis del dataset

* `codigo_alumno`
    - id
* `pais`
    - decision -> `eliminar la variable`
    - la mayoria es argentina 74.9%
    - 25% missing -> no esta mal
    - puedo dejarlo por casos particulares pero al no tener varianza no ayuda
* `localidad`
    - analizarlo mas en profundidad para tomar decision
* `provincia`
    - muy desbalanceado pero podria funcionar para algunas cosas
    - caba 46%, bs as 26%, missing 25%, y otros menores de 1%
* `titulo_secundario`
    - decisión -> `eliminar variable`
    - 51% de missing
    - se puede usar para profundizar algun análisis particular si vale la pena
* `estado_civil`
    - decision -> `eliminar variable`
    - muy desbalanceado
    - 92% soltero
    - se podria utilizar para algun analisis particular
    - dato no actualizado a través dle tiempo
* `sexo`
    - 83,4% M - 16,6% F
* `año_nacimiento`
    - decisión -> `usar variable`
    - tiene errores y outliers (nacimeinto en 1900 por ejemplo)
    - arreglar con criterio esos casos
    - modificar el nombre para sacar la letra ñ
* `año_ingreso`
    - decisión -> `usar variable`
    - hay casos viejo
        - el percentil 5 ya es 1999 de ingreso que podria ser un criterio para evitar problemas de outliers y fata de información en planes viejos y ese tipo de cosas
    - hay que analizar si las personas que año de ingreso viejos terminaron la facultad
* `codigo_propuesto`
    - decision -> `usar variable`
    - tener en cuenta que debe ser cuando ingresa peo tmb pudo haber cambaido de carrera
        - usar codigo de materia para saber eso
    - 9701 K, 3628 R, 3116 S, 2777 V, etc



## Datos Cursadas

curdf.columns

curdfpr = ProfileReport(curdf.drop(['codigo_alumno'], axis=1), title="Datos Cursadas - Profiling Report")

curdfpr.to_file("reportes/Cursadas_ProfileReport.html")

### Criterios basados en el análsis del dataset

* `codigo_alumno`
    - id
* `curso`
    - decisión -> `usar variable`
* `codigo_materia`
    - no soportado en el analisis estadistico
        - trabajar un poco en esta variable para dsps hacer estadistica
* `departamento`
    - decisión -> `usar variable`
    - a que departamento corresponde cada materia. Una carrera puede tener materias de distintos departamentos. como las básicas que las tienen todas las carreras
* `modalidad`
    - decisión -> `eliminar variable`
    - muy desbalanceada a precencial y las que no son solo presencial son un mix entre presencila y distancia
* `turno`
    - decisión -> `usar variable`
    - 10% missing -> ver si se puede averiguar los datos que faltan o tomar criterio
* `ciclo_lectivo`
    - decisión -> `usar variable`
    - el minimo es 1997
        - tener en cuenta esto por los datos viejos de año de ingreso o nacimiento
        - puede ser un criterio para descartar otros datos y hacer limpieza o modificar pensando que el dato de año viejo no es bueno
* `tipo_de_aprobacion`
    - decision -> `usarlo solo en caso de las promociones si es que no está el registro del final`
    - Regularidad, Promoción	289237	49.8%
    - Promoción, Regularidad	286618	49.3%
    - Regularidad	5424	 0.9%
    - Promoción 16	 < 0.1%
* `cantidad_cursada_regular`
    - decision -> `usar variable`
    - cantidad de veces que recurso una materia o el numero de vez que ese ese registro en el que esta recursando la materia
        - analizar un poco mas con los casos >0 a ver bien que significa.
        - esto puede ayudar a que si no tenemos el cuatrimestre pero el registro de la materia se duplica el ciclo_lectivo y este registro incrementa en u numero, identificaria el cuatrimestre
    - tiene muchos ceros pero eso significa que curso y aprobó y no tuvo necesidad de recursar
* `descripcion_recusada_regular`
    - decision -> `usar variable`
        - es la pareja de `cantidad_cursada_regular` -> `usar una de las 2 variables` o usarla para entender bien como es la variable
* `cantidad_recursada_libre`
    - decisión -> `usar variable`
        - tiene alta correlacion con la anterior `cantidad_cusada_regular`
        - analizar bien cual es la diferencia y como utilizarlas. quizas solo haya que utilizar una sola o transformar ambas en una sola variable
* `descripcion_recursada_libre`
    - decisión -> `usar variable`
        - es la pareja de `cantidad_recursada_libre`
        - la libreria d eprofiling no la soporto, analizar manualmente

---
* `WARNING`
    - hay varias filas duplicadas
        - ver bien porque y si es realidad esta mal calculado la variable de recursadas por ser el mismo ciclo lectivo o si se duplico sin querer y hay que borrarlo

## Datos Finales

findfpr = ProfileReport(findf.drop(['codigo_alumno'], axis=1), title="Datos Finales - Profiling Report")

findfpr.to_file("reportes/Finales_ProfileReport.html")

### Criterios basados en el análsis del dataset

* `codigo_alumno`
    - id
* `codigo_materia`
    - decision -> `usar variable`
    - no soportado
    - analizar manual o cambiarlo para que lo tome la libreria
* `anio_academico`
    - decisión -> `usar variable`
    - minimo 1972
* `nota`
    - decision -> `usar variable`
    - tiene mezclado numeros y 'aprobado' etc
        - hay que mejorar la variable para utilizarla -> hay otra variable `aprobado`. Hacer algo entre ambos.
        - tomar un criterio para esos casos
    - tiene missings
        - puede ser que sean ausentes
            - averiguar mejor analizando mas en profundidad los datos
            - adoptar criterio
* `aprobado`
    - decision -> `verificar`
    - tiene muchos missings
    - pareceria ser muy parecida a la variable `nota`
        - analizar bien la relacion entre ambas
        - quizas se pueda utilizar una sola
* `promociono`
    - decision -> `verificar` 
    - no soportada
        - analizar manualmente
* `año_cursada`
    - decision -> `usar variable pero verificar con el dataser de Cursadas`
    - minimo 1997
    - tiene 30% de missings

---
* `WARNING`
    - Tiene filas duplicadas
        - analizar bien que puede llegar a significar
            - si es un error -> borrarlas
            - intuision -> quizas si desaprobo o se ausento al final en el mismo año


# Cosas que habría que pedir de los datos

* `cursadas`
    - el cuatrimestre de la cursada
    - la o las notas de las cursadas
    - claridad sobre la variable `tipo_de_aprobacion`
