==> Notebooks/05_Limpieza_Datos_Finales.ipynb <==
# Datos Finales

import pandas as pd
import numpy as np

from ydata_profiling import ProfileReport

findf = pd.read_pickle('../data/interim/findf.pkl')
print('fin shape: ',findf.shape)

findf.columns

findf.head()

# Criterios de limpieza

## Criterios 

basados en el análsis del dataset hecho en el notebook 02_Analisis_Datos_Alumnos.ipynb

* `codigo_alumno`
    - id
* `codigo_materia`
    - decision -> `usar variable`
    - no soportado
    - analizar manual o cambiarlo para que lo tome la libreria
* `anio_academico`
    - decisión -> `usar variable`
    - minimo 1972
* `nota`
    - decision -> `usar variable`
    - tiene mezclado numeros y 'aprobado' etc
        - hay que mejorar la variable para utilizarla -> hay otra variable `aprobado`. Hacer algo entre ambos.
        - tomar un criterio para esos casos
    - tiene missings
        - puede ser que sean ausentes
            - averiguar mejor analizando mas en profundidad los datos
            - adoptar criterio
* `aprobado`
    - decision -> `verificar`
    - tiene muchos missings
    - pareceria ser muy parecida a la variable `nota`
        - analizar bien la relacion entre ambas
        - quizas se pueda utilizar una sola
* `promociono`
    - decision -> `verificar` 
    - no soportada
        - analizar manualmente
* `año_cursada`
    - decision -> `usar variable pero verificar con el dataser de Cursadas`
    - minimo 1997
    - tiene 30% de missings

---
* `WARNING`
    - Tiene filas duplicadas
        - analizar bien que puede llegar a significar
            - si es un error -> borrarlas
            - intuision -> quizas si desaprobo o se ausento al final en el mismo año


# Analisis por variables

## Variable: `codigo_materia`

findf.codigo_materia.describe()

#### Conclusion

hay mas valores unicos que en el dataset de cursadas. Por lo que hay finales que no estan en cursadas. Por lo que nos puede decir que falta informacion en cursadas.

se esperaria que fuera alreves.


hacer:

* traer dataset de cursadas y ver cuantos alumnos tienen finales pero no cursadas de esas materias. evaluar eliminacion o analizar porque. Puede ser que sean equivalencias aunque me parece raro que sean tantas.


pd.DataFrame(findf.codigo_materia.unique()).to_csv('../data/interim/materias_fin_tmp.csv', index=False)

findf.columns

## Variable: `anio_academico` y `año_cursada`

# pandas dataframe describe only for categorical variables
findf.describe(include=['object'])


findf.describe()

findf['anio_academico_str'] = findf['anio_academico'].astype(str)
findf['anio_cursada_str'] = findf['año_cursada'].astype(str)

aux = findf.groupby(['codigo_alumno', 'codigo_materia']).agg({'anio_academico_str': lambda x: x.nunique(), 'anio_cursada_str': lambda x: x.nunique()})
aux = aux.reset_index()

aux.columns = ['codigo_alumno', 'codigo_materia', 'anio_academico_str', 'anio_cursada_str']

aux[(aux.anio_cursada_str > 1)]

#### conclusion

no hay distntos valores en anio cursada

##### preguntas

el anio_cursada coincide con el maximo de anio_academicos?

aux = findf.groupby(['codigo_alumno', 'codigo_materia']).agg({
    'anio_academico_str': 'max', 
    'anio_cursada_str': 'max',
    'anio_academico': 'max', 
    'año_cursada': 'max'
    })
aux = aux.reset_index()

aux['iguales'] = aux.anio_academico_str == aux.anio_cursada_str
aux2 = aux[(aux.iguales == False)&(aux.anio_cursada_str != 'nan')&((aux.anio_academico) - (aux.año_cursada) > 1)].copy()

aux2.codigo_alumno.nunique()

#### conclusion

no se puede saber cual es la dif entre ambos años. Debido a que para un mismo alumno en todos los registros de finales es el mismo valor, se podria concluir cual es el año en que que se cursó la matria y aprobó. Se puede verificar.

curdf = pd.read_pickle('../data/interim/curdf.pkl')

curdf.columns

auxfin = findf[['codigo_alumno', 'codigo_materia', 'año_cursada']].drop_duplicates().copy()
# auxcur = curdf[['codigo_alumno', 'codigo_materia', 'ciclo_lectivo']].drop_duplicates().copy()
auxcur = curdf.groupby(['codigo_alumno', 'codigo_materia']).agg({'ciclo_lectivo':'max'}).reset_index().drop_duplicates().copy()

auxcurfin = pd.merge(auxfin, auxcur, how='left', on=['codigo_alumno', 'codigo_materia'])

auxcurfin.head()

auxcurfin['iguales'] = auxcurfin.año_cursada == auxcurfin.ciclo_lectivo

auxcurfin[(auxcurfin.iguales == False)&(auxcurfin.ciclo_lectivo.notnull())&(auxcurfin.año_cursada.notnull())]

auxcurfin[(auxcurfin.iguales == False)]['ciclo_lectivo'].describe()

auxcurfin[(auxcurfin.iguales == False)]['año_cursada'].describe()

auxcurfin.iguales.value_counts()

auxcurfin.shape

auxcurfin['NAN'] = auxcurfin['ciclo_lectivo'].isnull()

auxcurfin['NAN_iguales'] = auxcurfin['iguales'] == auxcurfin['NAN']
auxcurfin['NAN_iguales'].value_counts()

#### conclusiones

año cursada del dataset de finales es igual al del ciclo lectivo maximo agrupando por alumno+materia en todos los casos donde el ciclo lectivo no es nulo

hay datos solo en año cursada, que no estan en ciclo lectivo, por lo que puede ser que no esté el dato en cursadas pero si esta en finales y se podria tomar de ahi.

findf.columns

## Variables: `nota` y `aprobado` y `promociono`

findf.nota.unique()

findf.aprobado.unique()

findf.promociono.unique()

findf.head(20)

findf_notas = findf.copy()
findf_notas['nota_aprobo'] = findf.nota == findf.aprobado
findf_notas['nota_promociono'] = findf.nota == findf.promociono

findf_notas.head()

findf_notas[['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']].head(20)

findf_notas[(findf_notas.nota_aprobo == False)|(findf_notas.nota_promociono == False)][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[~((findf_notas.nota_aprobo == True)|(findf_notas.nota_promociono == True))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[((findf_notas.nota_aprobo == True)&(findf_notas.nota_promociono == True))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[((findf_notas.nota_aprobo == False))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[((findf_notas.nota.isnull())&findf_notas.aprobado.isnull()&findf_notas.promociono.isnull())][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

# nota distinto a aprobo y que aprobo no es nan
findf_notas[((findf_notas.nota != findf_notas.aprobado)&(findf_notas.aprobado.notnull()))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[((findf_notas.nota != findf_notas.promociono)&(findf_notas.promociono.notnull()))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[(((findf_notas.nota != findf_notas.promociono)&(findf_notas.nota.notnull()))&
             ((findf_notas.nota != findf_notas.promociono)&(findf_notas.promociono.notnull())))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

findf_notas[(((findf_notas.nota != findf_notas.aprobado)&(findf_notas.nota.notnull()))&
             ((findf_notas.nota != findf_notas.aprobado)&(findf_notas.aprobado.notnull())))][['nota', 'aprobado', 'promociono', 'nota_aprobo', 'nota_promociono']]

#### conclusiones

* `nota` son todas las notas pero no están las de `promocion`

* `aprobado` tiene las notas de `nota` excepto las notas de desaprobación numéricas, pero si tiene las notas de desaprobación de texto como 'No aprobado' 

* `promociono` tiene las notas de promoción numérica y 'Aprobado' y no esta ni en las variables `nota` ni `aprobado`

`criterios`: 

1. si en los 3 es 'NAN' se tomará como ausente

2. transformar las columnas de la siguiente manera:

2.a. tipo_aprobacion: aprobacion, acreditacion, promocion, ausente

2.b. nota_numerica: solo números. Si no hay número pero si hay 'aprobacion' o 'acreditacion' se tomará la nota mínima de aprobacion (4 o 7 porque cambió la reglamentación)

2.c. nota_cualitativa: aprobado, No aprobado, ausente, acreditado, no acreditado



findf.columns

findf.nota.unique()

findf['tipo_aprobacion'] = np.where(
                                (findf.nota == 'Acreditado')|(findf.nota == 'No Acredit'), 'Acreditacion',
                                        np.where(findf.nota.notnull(), 'Aprobacion',
                                                np.where(findf.promociono.notnull(), 'promocion', 'ausente')
                                                )
                                )

findf.tipo_aprobacion.unique()

findf['nota_comb'] = findf.nota.combine_first(findf.aprobado).combine_first(findf.promociono)
findf['nota_comb']

findf.nota_comb.unique()

num_map = {'10':10, '9':9, '8':8, '7':7, '6':6, '5':5, '4':4, '3':3, '2':2, '1':1, '0':0}
cual_map = {'Aprobada': 'Aprobada', 'No aprobad': 'No Aprobada', 'Acreditado': 'Acreditada', 'No Acredit': 'No Acreditada', 'No Suficie': 'No Aprobada'}

findf['nota_num'] = findf.nota_comb.map(num_map)
findf['nota_cual'] = findf.nota_comb.map(cual_map)

findf.nota_num.isnull().sum()

findf.nota_cual.isnull().sum()

findf.nota_num[0]


# tener en cuenta el cambio de nota de un año a otro porque hay que convertir las notas para que sean comparables
findf['nota_txt'] = np.where(findf.nota_cual.notnull(), findf.nota_cual, 
                            np.where(findf.nota_num >= 4, 'Aprobada', 'No Aprobada')
                            )

#### pendientes

hacer la transformacion de las notas numericas para igual las escalas luego de la reglamentacion que pasaron de 4 a 7 las aporbaciones d elos examenes


findf.columns

COLS = ['codigo_alumno', 'codigo_materia', 'anio_academico', 'año_cursada', 'tipo_aprobacion', 'nota_num', 'nota_cual']

findf[COLS].head()

findf.shape

findf.to_pickle('../data/interim/findf_01.pkl')
