==> Notebooks/06_Limpieza_Datasets.ipynb <==
# Limpieza de Todos los Datasets Iniciales

import pandas as pd
import numpy as np

from ydata_profiling import ProfileReport

curdf = pd.read_pickle('../data/interim/curdf.pkl')
print('cur shape: ',curdf.shape)
datdf = pd.read_pickle('../data/interim/datdf_clean.pkl')
print('datdf shape: ',datdf.shape)
findf = pd.read_pickle('../data/interim/findf_01.pkl')
print('fin shape: ',findf.shape)

## Alumnos a eliminar de los datasets

alumnos_delete = pd.read_pickle('../data/interim/alumnos_delete.pkl')
print('alumnos_delete shape: ',alumnos_delete.shape)

alumnos_delete.codigo_alumno.nunique()

## Datos de Alumnos

datdf.head()

# eliminamos los alumnos de alumnos_delete
datdf = datdf[~datdf.codigo_alumno.isin(alumnos_delete.codigo_alumno)]

datdf.codigo_alumno.nunique()

24830 - 22157

datdf.to_pickle('../data/raw_cleaned/datdf_01.pkl')
datdf.to_csv('../data/raw_cleaned/datdf_01.csv',index=False)

## Datos de Cursadas

curdf.head()

# eliminamos los alumnos de alumnos_delete
curdf = curdf[~curdf.codigo_alumno.isin(alumnos_delete.codigo_alumno)]

#### criterio

Por ahora nose sabe cual es la diferencia entre `cant_recursada_regular` y `cant_recursada_libre` por lo que voy a juntar ambas variables en solo cant de resursadas y que el número se incremente segun la cantidad de registros ordenado por ciclo lectivo.

curdf_aux = curdf.sort_values(['codigo_alumno', 'codigo_materia', 'ciclo_lectivo'], ascending=True).copy()

# agrupar y numero de fila
curdf_aux['cant_cursada'] = curdf.groupby(['codigo_alumno', 'codigo_materia'])[['ciclo_lectivo']].transform('rank', method='first')

curdf_aux['cant_total_cursadas'] = curdf.groupby(['codigo_alumno', 'codigo_materia'])[['ciclo_lectivo']].transform('count')
curdf_aux

### Limpieza de columnas




curdf_aux.columns

COLS = ['codigo_alumno', 'curso', 'codigo_materia', 'departamento', 'turno', 'ciclo_lectivo', 'cant_cursada', 'cant_total_cursadas']

curdf_aux = curdf_aux[COLS]

curdf_aux.head()

curdf_aux.codigo_alumno.nunique()

curdf = curdf_aux.copy()

curdf.to_pickle('../data/raw_cleaned/curdf_01.pkl')
curdf.to_csv('../data/raw_cleaned/curdf_01.csv',index=False)

## Dataset Cursadas complementario

curdf.codigo_materia.nunique()

findf.codigo_materia.nunique()

# eliminamos los alumnos de alumnos_delete
findf = findf[~findf.codigo_alumno.isin(alumnos_delete.codigo_alumno)]

findf.codigo_materia.nunique()

findf.codigo_alumno.nunique()

findf.columns

findf.head()

COLS = ['codigo_alumno', 'codigo_materia', 'anio_academico', 'nota_num', 'nota_cual', 'nota_txt', 'año_cursada',]

findf = findf[COLS]

findf.head()

findf = findf.sort_values(['codigo_alumno', 'codigo_materia', 'anio_academico', 'nota_num'], ascending=True).copy()

findf['cant_finales'] = findf.groupby(['codigo_alumno', 'codigo_materia'])[['anio_academico']].transform('rank', method='first')

findf.head()

findf.duplicated().sum()

## ACA esta el problema, tiene que ser por el par alumno manteria y no solo materia

finaux = findf.copy()
curaux = curdf.copy()

finaux['almat'] = finaux.codigo_alumno.astype(str) + '_' + finaux.codigo_materia.astype(str)
curaux['almat'] = curaux.codigo_alumno.astype(str) + '_' + curaux.codigo_materia.astype(str)

aux = finaux[~finaux.almat.isin(curaux.almat)].copy()

aux.codigo_materia.nunique()

# eliminamos los alumnos de alumnos_delete
#findf = findf[~findf.codigo_alumno.isin(alumnos_delete.codigo_alumno)]

aux.codigo_alumno.nunique()

### Criterio

no se pueden eliminar los alumnos que tienen codigo_materia en el datset de finales pero que no esta en cursadas porque nos quedamos sin datos

- se puede tomar el dato del año de cursada como el dato faltante de cuadno se cursó

aux.shape

aux.head()


# criterio
# si anio cursada no tiene datos, completar con el mismo año que rindio final que esta en anio_academico
aux['anio_cursada'] = aux['año_cursada'].combine_first(aux.anio_academico)
aux.head()

# me quedo con los datos para agregarlos al dataset de cursadas

COLS = ['codigo_alumno', 'codigo_materia', 'anio_cursada']
cursadas_criterio = aux[COLS].copy().drop_duplicates()
cursadas_criterio.shape

cursadas_criterio.head()

# este join no deberia tener interseccion
extra = pd.merge(cursadas_criterio, curdf, on=['codigo_alumno', 'codigo_materia'], how='inner')
extra.shape

curdf.columns

cursadas_criterio.columns

cursadas_criterio.columns = ['codigo_alumno', 'codigo_materia', 'ciclo_lectivo']

# completo el resto de las variables para dsps tener el dataset para integrarloa cursadas

cursadas_criterio['curso'] = 'S/D'
cursadas_criterio['departamento'] = 'S/D' # este se puede completar dsps con los datos de organigrama
cursadas_criterio['turno'] = 'S/D'
cursadas_criterio['cant_cursada'] = 1
cursadas_criterio['cant_total_cursadas'] = 1

cursadas_criterio = cursadas_criterio[curdf.columns]


cursadas_criterio.shape

cursadas_criterio.columns

# dataset de cursadas complementarias al original por datos obtenidos del dataset de Finales
cursadas_criterio.to_pickle('../data/raw_cleaned/curdf_complementario_01.pkl')
cursadas_criterio.to_csv('../data/raw_cleaned/curdf_complementario_01.csv', index=False)

## Dataset de Finales

findf.columns

findf = findf[['codigo_alumno', 'codigo_materia', 'anio_academico', 'nota_num', 'nota_cual', 'nota_txt', 'cant_finales']]

findf.to_pickle('../data/raw_cleaned/findf_01.pkl')
findf.to_csv('../data/raw_cleaned/findf_01.csv', index=False)
