==> Notebooks/20_Dataset_01.ipynb <==
# Universo / Target

## Definición de target

Target = 1 cuando el alumno no curso ni rindió finales por 2 ciclos lectivos consecutivos

## Dataset a armar

ejemplo

| Alumno | Ciclo | Target |
| --- | --- | --- |
| 1 | 2010 | 0 |
| 1 | 2011 | 0 |
| 1 | 2012 | 1 |
| 1 | 2013 | 1 |
| 2 | 2012 | 0 |
| 2 | 2013 | 0 |
| 2 | 2014 | 0 |
| 2 | 2015 | 0 |

- ciclo minimo es el año de ingreso del alumno
- ciclo maximo es el año de egreso del alumno o del ultimo registro (cursadas o finales) + 2




import pandas as pd
import numpy as np

from ydata_profiling import ProfileReport

## Datos de entrada

curdf = pd.read_pickle('../data/raw_cleaned/curdf_01.pkl')
print('cur shape: ',curdf.shape)
curdf_comp = pd.read_pickle('../data/raw_cleaned/curdf_complementario_01.pkl')
print('cur_comp shape: ',curdf_comp.shape)
datdf = pd.read_pickle('../data/raw_cleaned/datdf_01.pkl')
print('datdf shape: ',datdf.shape)
findf = pd.read_pickle('../data/raw_cleaned/findf_01.pkl')
print('fin shape: ',findf.shape)

## Armar Dataset con datos de cursadas (base + complementario) y finales

cur = pd.concat([curdf,curdf_comp],axis=0)
cur.head()

cur.shape

cur.tail()

findf.head()

df = pd.concat( 
    [cur[['codigo_alumno', 'ciclo_lectivo']].copy().drop_duplicates(), 
    findf[['codigo_alumno', 'anio_academico']].copy().drop_duplicates().rename(columns={'anio_academico':'ciclo_lectivo'})], axis=0). \
    drop_duplicates().sort_values(['codigo_alumno', 'ciclo_lectivo']).reset_index(drop=True)


df.head()

df.shape

df.duplicated().sum()

# traigo el año de ingreso para cada alumno

datdf.head()

aux = datdf[['codigo_alumno', 'anio_ingreso']].drop_duplicates().copy()

df = pd.merge(df, aux, on='codigo_alumno', how='left')

df.head()

df.shape

# calcular el tiempo que pasa entre un ciclo lectivo y el siguiente para un mismo alumno
aux = df.copy()
aux['delta_ciclo_lectivo'] = aux.groupby('codigo_alumno')['ciclo_lectivo'].transform(lambda x: x - x.shift(1))
aux.head(20)

(aux.delta_ciclo_lectivo > 2).value_counts()

3291/144245

### conclusión

tenemos un 2,28% de desertores

dataset desbalanceado


## Armar Dataset con datos de cursadas (base) y finales

df = pd.concat( 
    [curdf[['codigo_alumno', 'ciclo_lectivo']].copy().drop_duplicates(), 
    findf[['codigo_alumno', 'anio_academico']].copy().drop_duplicates().rename(columns={'anio_academico':'ciclo_lectivo'})], axis=0). \
    drop_duplicates().sort_values(['codigo_alumno', 'ciclo_lectivo']).reset_index(drop=True)


df.shape

# calcular el tiempo que pasa entre un ciclo lectivo y el siguiente para un mismo alumno
aux = df.copy()
aux['delta_ciclo_lectivo'] = aux.groupby('codigo_alumno')['ciclo_lectivo'].transform(lambda x: x - x.shift(1))
aux.head(20)

(aux.delta_ciclo_lectivo > 2).value_counts()

### conclusion

da la misma proporcion usando o no las cursadas complementarias

## Análsis de dataset si se hace a nivel alumno

### Contemplando solo el ultimo ciclo

# agrupar aux por alumno y filtrar fila con maximo ciclo lectivo
aux_alumno = aux.sort_values(['codigo_alumno', 'ciclo_lectivo']).groupby('codigo_alumno').last().reset_index().copy()

aux_alumno.head()


(aux_alumno.delta_ciclo_lectivo > 2).value_counts()

1238/20919

#### conclusión

si se hace a nivel alumno tenemos un desbalance del 6%


### Contemplando si alguna vez desertó a pensar de que despues retomó

# agrupar aux por alumno y quedarme con el máximo valor de delta_ciclo_lectivo
aux_alumno = aux.sort_values(['codigo_alumno', 'ciclo_lectivo']).groupby('codigo_alumno').max().reset_index().copy()

aux_alumno.head()


(aux_alumno.delta_ciclo_lectivo > 2).value_counts()

2769/19388

#### conclusión

- si se considera si desertó al menos una vez a pesar de que dsps retomó el desbalance es del 14,3%

# Nexts steps

1. por ahora no se contempla si un alumno regular abandona repentinamente la carrera xq al no haber registro siguiente, no se sabe si es un abandono y no se registra

2. en la condicion 1, hay q tener en cuenta que el alumno no haya terminado la carrera porque sería un falso desertor (en algunas carreras es obligatorio terminar con la materia proyecto final y en otras no. averiguar)

# Analisis de distribución de carreras

curdf.columns

datdf.columns

findf.columns

datdf.codigo_carrera_ingreso.value_counts()

curdf.curso.apply(lambda x: x[0]).value_counts()

curdf.columns

curdf[~curdf.departamento.duplicated()][['curso','departamento']]

## Decision

para no complicarse con tantas variantes, en primera instancia vamos a reducir a las carreras en las que se tenga datos de mas de 1000 alumnos y que ademas la ultima materia obligatoriamente sea proyecto final:

|cod|carrera|alumnos|
|---|----------|----------|
| K | sistemas |     8702 |
|R|electronica|     3278|
|S|mecanica     |2822|
|V|quimica     |2475|
|O|civil     |2045|
|Q|eléctrica|      928|
