==> Notebooks/21_Dataset_01.ipynb <==
# Universo / Target

## Definición de target

Target = 1 cuando el alumno no curso ni rindió finales por 2 ciclos lectivos consecutivos y no haya finalizado la carrera

## Dataset a armar

ejemplo

| Alumno | Ciclo | Target |
| --- | --- | --- |
| 1 | 2010 | 0 |
| 1 | 2011 | 0 |
| 1 | 2012 | 1 |
| 1 | 2013 | 1 |
| 2 | 2012 | 0 |
| 2 | 2013 | 0 |
| 2 | 2014 | 0 |
| 2 | 2015 | 0 |

- ciclo minimo es el año de ingreso del alumno
- ciclo maximo es el año de egreso del alumno o del ultimo registro (cursadas o finales) + 2




import pandas as pd
import numpy as np

from ydata_profiling import ProfileReport

## Datos de entrada

curdf = pd.read_pickle('../data/raw_cleaned/curdf_01.pkl')
print('cur shape: ',curdf.shape)
curdf_comp = pd.read_pickle('../data/raw_cleaned/curdf_complementario_01.pkl')
print('cur_comp shape: ',curdf_comp.shape)
datdf = pd.read_pickle('../data/raw_cleaned/datdf_01.pkl')
print('datdf shape: ',datdf.shape)
findf = pd.read_pickle('../data/raw_cleaned/findf_01.pkl')
print('fin shape: ',findf.shape)

## Reduccion de alumnos

- solo quedan los alumnos de las carreras:

|cod|carrera|alumnos|
|---|----------|----------|
| K | sistemas |     8702 |
|R|electronica|     3278|
|S|mecanica     |2822|
|V|quimica     |2475|
|O|civil     |2045|
|Q|eléctrica|      928|

datdf.columns

datdf_aux = datdf[datdf.codigo_carrera_ingreso.isin(['K','R','S','V','O','Q'])].copy()
datdf_aux.shape


curdf['dataset'] = 'cursadas'

curdf_aux = curdf[curdf.codigo_alumno.isin(datdf_aux.codigo_alumno)].copy()
curdf_aux.shape

curdf_comp['dataset'] = 'cursadas_complementario'

curdf_comp_aux = curdf_comp[curdf_comp.codigo_alumno.isin(datdf_aux.codigo_alumno)].copy()
curdf_comp_aux.shape

findf['dataset'] = 'finales'

findf_aux = findf[findf.codigo_alumno.isin(datdf_aux.codigo_alumno)].copy()
findf_aux.shape

### condición

- tener en cuenta la materia de proyecto final

Proyecto Final:

	- civil (O): 950289

	- eléctrica (Q): 950594

	- electrónica (R): 950459

	- mecánica (S): 940842

	- química (V): 951196

	- sistemas (K): 082037

### datasets

| codigo_alumno | anio_ingreso | ultima_cursada | ultimo_final | egreso_anio |
| --- | --- | --- | --- | --- |

CODIGOS_MATERIAS_PF = ['950289','950594','950459','940842','951196','082037']
CODIGOS_CARRERAS = ['O','Q','R','S','V','K']

codigos = pd.DataFrame.from_dict({'codigo_materia':CODIGOS_MATERIAS_PF,'codigo_carrera':CODIGOS_CARRERAS})
codigos

cur = pd.concat([curdf_aux,curdf_comp_aux],axis=0)
cur.head()

VARS = ['codigo_alumno','ciclo_lectivo','codigo_materia','codigo_carrera']

cur_pf = pd.merge(cur,codigos,on=['codigo_materia','codigo_materia'],how='inner')[VARS]
cur_pf.head()

cur_pf[['codigo_alumno','codigo_materia']].drop_duplicates().shape

print(cur_pf.codigo_alumno.shape)
print(cur_pf.codigo_alumno.nunique())

cur_pf_aux = cur_pf.groupby(['codigo_alumno','codigo_carrera'])['ciclo_lectivo'].max().reset_index().copy()
cur_pf_aux.head()

print(cur_pf_aux.codigo_alumno.shape)
print(cur_pf_aux.codigo_alumno.nunique())

#### conclusion

todos los alumnos que cursaron proyecto final lo hicieron en una sola carrera

findf_aux.columns

VARS = ['codigo_alumno', 'anio_academico', 'codigo_materia', 'codigo_carrera']

fin_pf = pd.merge(findf_aux,codigos,on=['codigo_materia','codigo_materia'],how='inner')[VARS]

print(fin_pf.shape)
print(fin_pf.codigo_alumno.nunique())

fin_pf_aux = fin_pf.groupby(['codigo_alumno','codigo_carrera'])['anio_academico'].max().reset_index().copy()
fin_pf_aux.head()

print(fin_pf_aux.shape)
print(fin_pf_aux.codigo_alumno.nunique())

#### conclusion

los alumnos que rindieron final lo hicieron para una carrera sola

fin_pf.codigo_alumno.duplicated().value_counts()

aux = fin_pf[fin_pf.codigo_alumno.isin(fin_pf[fin_pf.codigo_alumno.duplicated()]['codigo_alumno'])].copy()

pd.merge(aux, findf, on=['codigo_alumno','codigo_materia', 'anio_academico'], how='inner').drop_duplicates()

#### conclusion

los finales duplicados son ausentes

## Armado del dataset

| codigo_alumno | anio_ingreso | ultima_cursada | ultimo_final | egreso_anio |
| --- | --- | --- | --- | --- |

datdf_aux.columns

print(datdf_aux.shape)
print(datdf_aux.codigo_alumno.nunique())

# aux1 = datdf[['codigo_alumno','anio_ingreso']].copy()

cur_pf_aux.columns

aux2 = cur_pf_aux[['codigo_alumno','ciclo_lectivo']].copy()

fin_pf_aux.columns

aux3 = fin_pf_aux[['codigo_alumno','anio_academico']].copy()

# juntamos los 3 datasets

df_pf = pd.merge(aux2, aux3, on='codigo_alumno', how = 'left' )

df_pf = pd.merge(df_pf, datdf_aux, on = 'codigo_alumno', how = 'inner' )

df_pf.head()

df_pf.codigo_alumno.nunique()

## Nexts Steps

- tengo el dataset de ultimas cursadas y ultimos finales dados
---
- [ ] TODO: TARGET. falta armar el dataset con el target. 
    - tener en cuenta el año de la cursada + 2 para q sea desertor
    - si rindio el final, egreso y no deberia contemplarse como desertor a pesar de haberse recibido antes

---
- [ ] TODO: DATASET
    - 1. a nivel alumno
    - 2. a nivel ciclo lectivo para hacerlo temporal

findf_aux.columns

findf_aux.shape

findf_aux.codigo_alumno.nunique()

# ultimo final de los alumnos
aux = findf_aux.groupby(['codigo_alumno'])['anio_academico'].max().reset_index().copy()
aux.columns = ['codigo_alumno','ultimo_final']
aux.head()

# agrego la fecha que rindió proyecto final
aux = pd.merge(aux, fin_pf_aux[['codigo_alumno', 'anio_academico']], on='codigo_alumno', how='left')
aux.columns = ['codigo_alumno', 'ultimo_final', 'final_proyecto_final']
# agrego si el final fue aoprobado para saber si se recibió
aux.head()

fin_pf.shape

fin_pf_aux.shape

fin_pf.columns

aux_pf = pd.merge(fin_pf, findf, on=['codigo_alumno','codigo_materia', 'anio_academico'], how='inner').drop_duplicates()
aux_pf = aux_pf[aux_pf.nota_txt == 'Aprobada']

aux_pf.head()

fin_pf_aux.shape

fin_pf_aux.head()

aux_pf2 = pd.merge(fin_pf_aux, aux_pf, on=['codigo_alumno','anio_academico'], how = 'left')

aux_pf2.nota_txt.value_counts()

fin_pf[ fin_pf.codigo_alumno.isin( aux_pf2[aux_pf2.nota_txt.isna()]['codigo_alumno'])]

revisar codigo anterior pero de todos los finales hay 2 personas que tienen ausente y no se presentaron nuevamente. raro xq es del año 2020 y 2021

aux


aux_pf2

aux = pd.merge(aux, aux_pf2[['codigo_alumno', 'nota_txt']], on='codigo_alumno', how='left')
aux.columns = ['codigo_alumno', 'ultimo_final', 'final_proyecto_final', 'nota_txt_proyecto_final']

aux

cur.columns

cur_aux = cur.groupby(['codigo_alumno'])['ciclo_lectivo'].max().reset_index().copy()
cur_aux.shape

cur_aux

# agrego la ultima cursada de cada alumno
aux = pd.merge(aux, cur_aux, on='codigo_alumno', how='left')
aux

aux.columns = ['codigo_alumno', 'ultimo_final', 'final_proyecto_final', 'nota_txt_proyecto_final', 'ultima_cursada']

aux = aux[['codigo_alumno', 'ultima_cursada', 'ultimo_final', 'final_proyecto_final', 'nota_txt_proyecto_final']]

aux

datdf_aux.columns


datdf_aux.shape

aux = pd.merge(aux, datdf_aux[['codigo_alumno','anio_ingreso']], on='codigo_alumno', how='left')
aux

aux = pd.merge(aux, datdf_aux[['codigo_alumno','codigo_carrera_ingreso']], on='codigo_alumno', how='left')
aux = pd.merge(aux, cur_pf_aux[['codigo_alumno','codigo_carrera']], on='codigo_alumno', how='left')

aux

print((aux.codigo_carrera_ingreso == aux.codigo_carrera).sum())

print((~aux.codigo_carrera.isna()).sum())

aux[(~aux.codigo_carrera.isna())&(aux.codigo_carrera_ingreso != aux.codigo_carrera)]

#### conclusiones

- hay 5 personas que cambiaron de carrera y la finalizaron

cur_pf_aux.codigo_alumno.nunique()

aux[(~aux.nota_txt_proyecto_final.isna())&(aux.codigo_carrera_ingreso != aux.codigo_carrera)]

(~aux.codigo_carrera.isna()).sum()

aux.shape

aux

aux.columns


aux = aux [['codigo_alumno', 'anio_ingreso', 'codigo_carrera_ingreso', 'ultima_cursada', 'ultimo_final', 'final_proyecto_final', 'nota_txt_proyecto_final', 'codigo_carrera']]
aux.columns = ['codigo_alumno', 'anio_ingreso', 'codigo_carrera_ingreso', 'ultima_cursada', 'ultimo_final', 'final_proyecto_final', 'nota_txt_proyecto_final', 'codigo_fin_carrera']

aux

aux.nota_txt_proyecto_final.value_counts()

1513 / 20250

aux.to_pickle('../data/processed/dataset_intermedio_primeros_y_ultimos_registros.pkl')
aux.to_csv('../data/processed/dataset_intermedio_primeros_y_ultimos_registros.csv', index=False)

datdf_aux.to_pickle('../data/raw_cleaned/datdf_02.pkl')
datdf_aux.to_csv('../data/raw_cleaned/datdf_02.csv', index=False)
curdf_aux.to_pickle('../data/raw_cleaned/curdf_02.pkl')
curdf_aux.to_csv('../data/raw_cleaned/curdf_02.csv', index=False)
findf_aux.to_pickle('../data/raw_cleaned/findf_02.pkl')
findf_aux.to_csv('../data/raw_cleaned/findf_02.csv', index=False)


curdf_comp_aux.to_pickle('../data/raw_cleaned/curdf_complementario_02.pkl')
curdf_comp_aux.to_csv('../data/raw_cleaned/curdf_complementario_02.csv', index=False)

### Dataset intermedio con datos de Proyecto Final y ultima cursada

- este dataset es un dataset intermedio que ayudará para armar los datasets para los modelos

1. sabemos quienes egresaron y en que año para no tomarlos como desertores a pesar de que no haya actividad
2. sabemos cual es la ultima cursada y/o ultimo final sobre el cual hay que sumarle 2 años para considerarlo como desertor (si no se recibió - caso 1.)
3. sabemos a partir de que año (anio_ingreso) hay que armar para cada alumno o contar para cada alumno que es lo que pasa en cada año para poder hacer un corte en cualquier año y hacer datasets temporales
