==> Notebooks/23_Dataset_01.ipynb <==
# Universo / Target

## Definición de target

Target = 1 cuando el alumno no curso ni rindió finales por 2 ciclos lectivos consecutivos y no haya finalizado la carrera

## Dataset a armar

ejemplo

| Alumno | Ciclo | Target |
| --- | --- | --- |
| 1 | 2010 | 0 |
| 1 | 2011 | 0 |
| 1 | 2012 | 1 |
| 1 | 2013 | 1 |
| 2 | 2012 | 0 |
| 2 | 2013 | 0 |
| 2 | 2014 | 0 |
| 2 | 2015 | 0 |

- ciclo minimo es el año de ingreso del alumno
- ciclo maximo es el año de egreso del alumno o del ultimo registro (cursadas o finales) + 2




import pandas as pd
import numpy as np

import seaborn as sns

from ydata_profiling import ProfileReport

## Datos de entrada

cur = pd.read_pickle('../data/raw_cleaned/curdf_03.pkl')
fin = pd.read_pickle('../data/raw_cleaned/findf_03.pkl')
curcomp = pd.read_pickle('../data/raw_cleaned/curdf_complementario_03.pkl')
dat = pd.read_pickle('../data/raw_cleaned/datdf_03.pkl')
df_ult = pd.read_pickle('../data/processed/dataset_intermedio_primeros_y_ultimos_registros_03.pkl')
df = pd.read_pickle('../data/processed/df_agrupado_target_03.pkl')


print('cur shape: ',cur.shape)
print('cur_comp shape: ',curcomp.shape)
print('datdf shape: ',dat.shape)
print('fin shape: ',fin.shape)
print('df_ultimos_registros shape: ',df_ult.shape)
print('df_agrupado_target shape: ',df.shape)

### Planificacion de Datasets

cur.head()

curcomp.head()

dat.head()

fin.head()

# contar valores dentro del groupby 
aux = fin.groupby(['codigo_alumno','codigo_materia'])['nota_txt'].value_counts().reset_index(name='count')
aux = aux[(aux['count'] > 1)&(aux.nota_txt == 'Aprobada')].copy()
aux

aux2 = pd.merge(aux,fin, on=['codigo_alumno','codigo_materia','nota_txt'], how='left')

### criterio

- se encontraron datos como que los alumnos se sacaban un 10 y despues de 2 años rendiand de nuevo y se sacanab  un 4. ademas de ser absurdo y debido a la cantidad de errores y sin forma de poder corregirlo por ahora el proyecto queda pausado hasta que se pueda corregir los datos
